Use predefined function types for parse driver functions used to set up function pointers.
Account for changed prototype of parse_inp_fnc_t functions.

Cast parse conversion results to appropriate types to avoid compiler warnings.

Let ioctl() for Windows accept a (void *) to avoid compiler warnings
when called with pointers to different types.

+++ libparse/gpstolfp.c
@@ -50,7 +50,7 @@ gpstolfp(
       weeks += GPSWEEKS;
     }
 
-  lfp->l_ui = weeks * SECSPERWEEK + days * SECSPERDAY + seconds + GPSORIGIN; /* convert to NTP time */
+  lfp->l_ui = (uint32_t)(weeks * SECSPERWEEK + days * SECSPERDAY + seconds + GPSORIGIN); /* convert to NTP time */
   lfp->l_uf = 0;
 }
